chore: rename data table heading file
