--- packages/core/src/components/cv-data-table/cv-data-table-heading-label.vue
@@ -1,11 +1,0 @@
-<template>
-  <span class="bx--table-header-label">
-    <slot />
-  </span>
-</template>
-
-<script>
-export default {
-  name: 'CvDataTableHeadingLabel',
-};
-</script>
