+++ packages/core/src/components/cv-data-table/cv-data-table-heading.vue
