+++ packages/core/src/components/cv-data-table/index.js
@@ -2,7 +2,7 @@ import CvDataTableAction from './cv-data-table-action';
 import CvDataTableCell from './cv-data-table-cell';
 import CvDataTableRow from './cv-data-table-row';
 import CvDataTable from './cv-data-table';
-import CvDataTableHeading from './_cv-data-table-heading';
+import CvDataTableHeading from './cv-data-table-heading';
 import CvDataTableSkeleton from './cv-data-table-skeleton';
 
 export { CvDataTable, CvDataTableRow, CvDataTableCell, CvDataTableAction, CvDataTableHeading, CvDataTableSkeleton };
